requirements.txt
future
emoji==1.7.0
lxml
wget
bs4
gtts
beautifulsoup4==4.11.1
requests
python-telegram-bot==13.14
SQLAlchemy==1.3.20
psycopg2-binary
feedparser
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
jikanpy
speedtest-cli==2.1.3
coffeehouse
regex
bleach
wikipedia
markdown2
telethon==1.25.2
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
pyrogram==2.0.57
PyDictionary
git+https://github.com/Sorrybabe/downloader
youtube_search_python==1.6.6
youtube_search
lyrics_extractor
asyncio
dateparser==1.1.1
pymongo
dnspython
secureme
apscheduler
html2text
bs4
redis
pySmartDL
search_engine_parser
pytz
jikanpy
tswift
python-arq
envparse
textblob
cloudscraper
gpytranslate
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
